relocate font util into separate util dir
